refactored new util file and added comments
