+++ src/actions/attendees/index.js
@@ -1,3 +1,4 @@
 export * from './setAttendeesByTicket'
 export * from './setRestrictedAttendeeIds'
 export * from './initSortedAttendees'
+export * from './initRestrictedAttendees'
